kernels: move importing-related functions to `importer`
Ideally we'd call the module `import`, but it's a reserved keyword.
